RD: modify timer timing rig
Modify the timing rig to track the average delta between two
subsequent timer callbacks. The original rigs track the absolute
error versus nominal clock, but this not what we want to track.
If a timer cb is late, we expect
  - the next timer cb to come early (so avg stays the same)
  - we capture the event via worst-case metric

Expected result is avg is +-tick to the nomimal value and a small
worst-case value.

Another change is to add ALL_CORES_BUSY define to test with and without
all CPUs spinning during test. The behaviour seems to be different when
CPUs are not fully loaded by the worker threads.